Publisher - move TaskPublishGo.py to TaskPublish.py
